relocation all from app.component in service
